+++ python/samples/documentation_examples/plugins/WriterPlugin/ShortPoem/config.json
@@ -4,7 +4,7 @@
   "execution_settings": {
     "default": {
       "max_tokens": 200,
-      "temperature": 0.5,
+      "temperature": 0.7,
       "top_p": 0.0,
       "presence_penalty": 0.0,
       "frequency_penalty": 0.0
